Fixed/improved perturbation sampling

+++ src/core/lib/lattice/field2n.cpp
@@ -74,7 +74,7 @@ Field2n::Field2n(const DCRTPoly & DCRTelement)
 }
 
 //Constructor from a ring element matrix
-Field2n::Field2n(const Matrix<int32_t> &element)
+Field2n::Field2n(const Matrix<int64_t> &element)
 {
 	for (size_t i = 0; i < element.GetRows(); i++) {
 		this->push_back(element(i, 0));
